chore: only start scheduler in Flask reloader thread
Our python template uses Flask as the dev server which has a built-in
reloader mechanism so that Flask restarts when it detects a file change.
This uses separate processes: main, which always runs and child, which
gets restarted.

The two processes start the scheduler which makes us send the syncing
request twice. This only happens in DEV mode though. This change ensures
that we don't start the scheduler in the main process, so that only one
is running at a time.

I'm not sure if this code works in production mode yet, but I'm
committing so it's definitely stored in the git history and we don't
lose it if we ever want to fix this issue here or more generally in the
template.

!REVERT THIS UNLESS TESTED PROPERLY IN PROD MODE!