chore: remove unused endpoints
Most of these are for setting or getting related data (signers/piece of
a signflow), but we don't use them and don't tend to work with them in
the stack. We could come back to these and re-add them, but that should
require a broader discussion about how we want to interact with our
services from the frontend.