feat: rewrite remove_signflows query to use UNIONs
The previous query used a lot of OPTIONALs which was too heavy for
Virtuoso. The usage of many variable names (all the combinations of
?thing_predicate ?thing_object) also made it far too heavy for
mu-authorization (as mu-authorization first does a SELECT query to
verify that the user has read-rights for those types, and with that many
variables it would always fail).