chore: use latest image of template and disable SPARQL logs
The latest untagged version of the template provides environment
variables to control whether SPARQL queries are logged. Additionally,
this version of the template logs SPARQL queries when an error is encountered.

Here we disable logging of SPARQL queries & by using the latest version we
ensure failed queries are still logged.