fix: don't try to delete files before user removes sign flow
If a user were to call this endpoint without being authorized for writing in the signing
graph, they would be able to delete the physical files (as that operation is completely
unrelated to mu-auth thus completely bypassing the authorization system of the
stack).

By first performing the update query, we ensure the user has the correct rights.